Handling erros middleware added to api

+++ Bookshelf/Bookshelf.Application/Features/Books/Commands/DeleteBook/DeleteBookCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using Bookshelf.Application.Contracts.Persistence;
+using Bookshelf.Application.Exceptions;
+using Bookshelf.Domain.Entities;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -15,6 +17,12 @@ namespace Bookshelf.Application.Features.Books.Commands.DeleteBook
         public async Task<Unit> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
         {
             var bookToDelete = await _bookRepository.GetByIdAsync(request.Id);
+
+            if (bookToDelete == null)
+            {
+                throw new NotFoundException(nameof(Book), request.Id);
+            }
+
             await _bookRepository.DeleteAsync(bookToDelete);
 
             return Unit.Value;
