+++ Bookshelf/Bookshelf.Application/Features/Categories/Commands/DeleteCategory/DeleteCategoryCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using Bookshelf.Application.Contracts.Persistence;
+using Bookshelf.Application.Exceptions;
+using Bookshelf.Domain.Entities;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -15,6 +17,12 @@ namespace Bookshelf.Application.Features.Categories.Commands.DeleteCategory
         public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
             var categoryToDelete = await _categoryRepository.GetByIdAsync(request.Id);
+
+            if (categoryToDelete == null)
+            {
+                throw new NotFoundException(nameof(Category), request.Id);
+            }
+
             await _categoryRepository.DeleteAsync(categoryToDelete);
 
             return Unit.Value;
