+++ Bookshelf/Bookshelf.Application/Features/Shelves/Commands/DeleteShelf/DeleteShelfCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using Bookshelf.Application.Contracts.Persistence;
+using Bookshelf.Application.Exceptions;
+using Bookshelf.Domain.Entities;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -15,6 +17,12 @@ namespace Bookshelf.Application.Features.Shelves.Commands.DeleteShelf
         public async Task<Unit> Handle(DeleteShelfCommand request, CancellationToken cancellationToken)
         {
             var shelfToDelete = await _shelfRepository.GetByIdAsync(request.Id);
+
+            if (shelfToDelete == null)
+            {
+                throw new NotFoundException(nameof(Shelf), request.Id);
+            }
+
             await _shelfRepository.RemoveAllShelfBooks(request.Id);
             await _shelfRepository.DeleteAsync(shelfToDelete);
             return Unit.Value;
